# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * todo = for today, pending = carried over, complete = done
 */
export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus];

export const TicketStatus = {
  todo: "todo",
  pending: "pending",
  complete: "complete",
} as const;

export interface Ticket {
  id: number;
  title: string;
  description: string;
  /** US state name (e.g. Virginia, Texas) */
  state: string;
  /** Name of the person who submitted the ticket */
  submitter: string;
  /** Work category (e.g. Power Automate, ArcGIS, CRM, Dashboard) */
  category: string;
  /** todo = for today, pending = carried over, complete = done */
  status: TicketStatus;
  /** The date this ticket was marked pending for */
  pendingDate?: string | null;
  /** When the ticket was completed */
  completedAt?: string | null;
  /** When the ticket was submitted */
  submittedAt: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateTicketRequestStatus =
  (typeof CreateTicketRequestStatus)[keyof typeof CreateTicketRequestStatus];

export const CreateTicketRequestStatus = {
  todo: "todo",
  pending: "pending",
  complete: "complete",
} as const;

export interface CreateTicketRequest {
  title: string;
  description: string;
  state: string;
  submitter: string;
  category: string;
  status?: CreateTicketRequestStatus;
  submittedAt?: string;
}

export type UpdateTicketRequestStatus =
  (typeof UpdateTicketRequestStatus)[keyof typeof UpdateTicketRequestStatus];

export const UpdateTicketRequestStatus = {
  todo: "todo",
  pending: "pending",
  complete: "complete",
} as const;

export interface UpdateTicketRequest {
  title?: string;
  description?: string;
  state?: string;
  submitter?: string;
  category?: string;
  status?: UpdateTicketRequestStatus;
  pendingDate?: string | null;
  completedAt?: string | null;
  submittedAt?: string;
}

export interface ParseImageRequest {
  /** Base64-encoded image data (with or without data URI prefix) */
  imageBase64: string;
  /** Image MIME type (e.g. image/png, image/jpeg) */
  mimeType?: string;
}

export interface ParseImageResponse {
  title: string | null;
  description: string | null;
  submitter: string | null;
  state: string | null;
  category: string | null;
  /** Brief note about extraction confidence */
  confidence: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type ListTicketsParams = {
  state?: string;
  submitter?: string;
  status?: string;
  category?: string;
};