# lib/api-zod/src/generated/types/parseImageRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ParseImageRequest {
  /** Base64-encoded image data (with or without data URI prefix) */
  imageBase64: string;
  /** Image MIME type (e.g. image/png, image/jpeg) */
  mimeType?: string;
}