# lib/api-zod/src/generated/types/parseImageResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Project Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ParseImageResponse {
  title: string | null;
  description: string | null;
  submitter: string | null;
  state: string | null;
  category: string | null;
  /** Brief note about extraction confidence */
  confidence: string | null;
}